Moved MenuConfigurator to util
Also renamed it to js instead of jsx
