+++ client/src/util/MenuConfigurator.js
